--- Assets/Scripts/UI/UIManager.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 24350d3971d366c4cb01fd042798d94a
-timeCreated: 1474990583
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
