+++ Assets/_SCENES/MAIN_MENU.unity
@@ -340,7 +340,6 @@ GameObject:
   m_Component:
   - 4: {fileID: 1063354127}
   - 114: {fileID: 1063354128}
-  - 114: {fileID: 1063354129}
   m_Layer: 0
   m_Name: GameController
   m_TagString: Untagged
@@ -373,17 +372,6 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   UIManager: {fileID: 0}
---- !u!114 &1063354129
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_PrefabParentObject: {fileID: 0}
-  m_PrefabInternal: {fileID: 0}
-  m_GameObject: {fileID: 1063354126}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 5e7a894cc58fce74f8f9919998cfdc22, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
 --- !u!1 &1156132827
 GameObject:
   m_ObjectHideFlags: 0
